requirements.txt
numpy
matplotlib
tensorflow
scikit-learn
pillow
opencv-python
streamlit